Porting to 64bit platform

+++ src/SALOMELocalTrace/Test/SALOMELocalTraceTest.cxx
@@ -180,9 +180,9 @@ SALOMELocalTraceTest::testLoadBufferPoolFile()
 
 void *PrintHello(void *threadid)
 {
-  int id_thread = (int)threadid;
+  long id_thread = (long)threadid;
   for (int i=0; i<NUM_MESSAGES;i++)
-    MESSAGE("Hello World! This is a trace test : " << id_thread 
-	    << " - iter " << i);
+  MESSAGE("Hello World! This is a trace test : " << id_thread 
+        << " - iter " << i);
   pthread_exit(NULL);
 }
